# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Waste Platform API",
    "description": "منصة إدارة النفايات الصلبة وإعادة التدوير — REST API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/admin/waste-types": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "category",
                  "nameAr",
                  "nameEn",
                  "unit",
                  "pricePerUnit"
                ],
                "properties": {
                  "code": {
                    "minLength": 2,
                    "maxLength": 32,
                    "pattern": "^[A-Z0-9_]+$",
                    "type": "string"
                  },
                  "category": {
                    "minLength": 2,
                    "maxLength": 32,
                    "type": "string"
                  },
                  "nameAr": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "nameEn": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "unit": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "bottle"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "liter"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "kg"
                        ]
                      }
                    ]
                  },
                  "pricePerUnit": {
                    "pattern": "^\\d+(\\.\\d{1,2})?$",
                    "type": "string"
                  },
                  "capacityWeightKg": {
                    "pattern": "^\\d+(\\.\\d{1,3})?$",
                    "type": "string"
                  },
                  "minWeightKg": {
                    "pattern": "^\\d+(\\.\\d{1,3})?$",
                    "type": "string"
                  },
                  "isBulkOnly": {
                    "type": "boolean"
                  },
                  "displayOrder": {
                    "minimum": 0,
                    "maximum": 999,
                    "type": "integer"
                  },
                  "referencePricePerTon": {
                    "pattern": "^\\d+(\\.\\d{1,2})?$",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/waste-types/{code}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "minLength": 2,
                    "maxLength": 32,
                    "pattern": "^[A-Z0-9_]+$",
                    "type": "string"
                  },
                  "category": {
                    "minLength": 2,
                    "maxLength": 32,
                    "type": "string"
                  },
                  "nameAr": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "nameEn": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "unit": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "bottle"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "liter"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "kg"
                        ]
                      }
                    ]
                  },
                  "pricePerUnit": {
                    "pattern": "^\\d+(\\.\\d{1,2})?$",
                    "type": "string"
                  },
                  "capacityWeightKg": {
                    "pattern": "^\\d+(\\.\\d{1,3})?$",
                    "type": "string"
                  },
                  "minWeightKg": {
                    "pattern": "^\\d+(\\.\\d{1,3})?$",
                    "type": "string"
                  },
                  "isBulkOnly": {
                    "type": "boolean"
                  },
                  "displayOrder": {
                    "minimum": 0,
                    "maximum": 999,
                    "type": "integer"
                  },
                  "referencePricePerTon": {
                    "pattern": "^\\d+(\\.\\d{1,2})?$",
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "code",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/addons": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "nameAr",
                  "nameEn",
                  "bonusPercent",
                  "appliesTo"
                ],
                "properties": {
                  "code": {
                    "minLength": 2,
                    "maxLength": 32,
                    "pattern": "^[A-Z_]+$",
                    "type": "string"
                  },
                  "nameAr": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "nameEn": {
                    "minLength": 2,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "bonusPercent": {
                    "pattern": "^\\d+(\\.\\d{1,2})?$",
                    "type": "string"
                  },
                  "appliesTo": {
                    "minItems": 1,
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "maxLength": 32,
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/admin/addons/{code}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "appliesTo"
                ],
                "properties": {
                  "appliesTo": {
                    "minItems": 1,
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "maxLength": 32,
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "code",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}